@@ -1,6 +1,3 @@
-// const formatters = require('./../lib/formatters');
-'use strict';
-
-module.exports = async ({debug}) => {
+export default async ({ debug }) => {
   debug('hello there!');
 };
