@@ -1,7 +1,5 @@
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
+import fs from 'node:fs';
+import path from 'node:path';
 
 /*
  * TBD
@@ -57,7 +55,7 @@ class App {
     this.product = product || config.get('system.product') || 'lando';
 
     // get needed props from the system config for other
-    const {cacheDir, configDir, dataDir, instance} = config.get('system');
+    const { cacheDir, configDir, dataDir, instance } = config.get('system');
 
     // set other props that are name-dependent
     this.cacheDir = path.resolve(cacheDir, '..', `${this.product}.${this.id}`);
@@ -70,7 +68,7 @@ class App {
 
     // created needed dirs
     for (const dir of [this.cacheDir, this.configDir, this.dataDir, this.logsDir, this.syscacheDir]) {
-      fs.mkdirSync(path.dirname(dir), {recursive: true});
+      fs.mkdirSync(path.dirname(dir), { recursive: true });
       this.debug('ensured directory %o exists', dir);
     }
 
@@ -84,9 +82,9 @@ class App {
       'logs-dir': this.logsDir,
       'plugins-dir': [...appConfig.get('plugin-dirs'), this.pluginsDir].filter(Boolean),
     };
-    this.config = config.newConfig({id: this.name, cached: path.join(this.cacheDir, 'config.json')});
-    this.config.add('app', {type: 'literal', store: {app: appStuff, ...this.appConfig.getUncoded('config')}});
-    this.config.add(this.product, {type: 'literal', store: config.getUncoded()});
+    this.config = config.newConfig({ id: this.name, cached: path.join(this.cacheDir, 'config.json') });
+    this.config.add('app', { type: 'literal', store: { app: appStuff, ...this.appConfig.getUncoded('config') } });
+    this.config.add(this.product, { type: 'literal', store: config.getUncoded() });
     // dump the cache
     this.config.dump();
 
@@ -94,7 +92,7 @@ class App {
     // we do this because we need a way a to cache things before plugins/the registry are compiled,
     // because we cannot use something like async getComponent inside the constructor and because we do not
     // want a full-blow async init() method we need to call EVERY time we new Bootstrap()
-    this.#_cache = new StorageBackend(({debug: this.debug.extend('#appcache'), dir: this.syscacheDir}));
+    this.#_cache = new StorageBackend({ debug: this.debug.extend('#appcache'), dir: this.syscacheDir });
 
     // if no-cache is set then lets force a cache wipe
     // if caching is disable and cache dir exists then flush
@@ -109,9 +107,11 @@ class App {
 
     // pseudo protected props
     this._componentsCache = {};
-    this._componentAlises = Object.fromEntries(Object.entries(this.config.getUncoded('core'))
-      .filter(([type, component]) => typeof component === 'string')
-      .map(([type, component]) => ([`core.${type}`, `${type}.${component}`])));
+    this._componentAlises = Object.fromEntries(
+      Object.entries(this.config.getUncoded('core'))
+        .filter(([, component]) => typeof component === 'string')
+        .map(([type, component]) => [`core.${type}`, `${type}.${component}`]),
+    );
   }
 
   // @TODO: the point of this is to have a high level way to "fetch" a certain kind of plugin eg global and
@@ -127,7 +127,7 @@ class App {
     const request = require('../utils/parse-package-name')(name);
     const source = request.peg ? request.peg : `^${plugin.version}`;
     // modify the landofile with the updated plugin
-    this.appConfig.save({plugins: {[plugin.name]: source}});
+    this.appConfig.save({ plugins: { [plugin.name]: source } });
 
     // reinit
     this.#_reint();
@@ -140,7 +140,7 @@ class App {
   getComponent(component) {
     return require('../utils/get-component')(
       component,
-      new this.Config({data: this.getRegistry(), id: `${this.id}-component-registry`}),
+      new this.Config({ data: this.getRegistry(), id: `${this.id}-component-registry` }),
       {
         aliases: this._componentAlises,
         cache: this._componentsCache,
@@ -151,26 +151,20 @@ class App {
   }
 
   // helper to get a component (and config?) from the registry
-  async getComponentInstance(component, constructor = {}, config = this.config, opts = {}) {
-    return require('../utils/get-component-instance')(
-      component,
-      constructor,
-      {
-        aliases: this._componentAlises,
-        cache: this._componentsCache,
-        config: this.config.get(),
-        debug: this.debug.extend('get-component-instance'),
-        registry: new this.Config({data: this.getRegistry(), id: `${this.id}-component-registry`}),
-      },
-    );
+  async getComponentInstance(component, constructor = {}) {
+    return require('../utils/get-component-instance')(component, constructor, {
+      aliases: this._componentAlises,
+      cache: this._componentsCache,
+      config: this.config.get(),
+      debug: this.debug.extend('get-component-instance'),
+      registry: new this.Config({ data: this.getRegistry(), id: `${this.id}-component-registry` }),
+    });
   }
 
   getHooks() {
-    const hooks = this.config.merge(this.hooks,
-      [
-        this.manifest.getUncoded('hooks.product', {ams: 'aoa'}),
-        this.manifest.getUncoded(`hooks.${this.name}`, {ams: 'aoa'}),
-      ],
+    const hooks = this.config.merge(
+      this.hooks,
+      [this.manifest.getUncoded('hooks.product', { ams: 'aoa' }), this.manifest.getUncoded(`hooks.${this.name}`, { ams: 'aoa' })],
       ['aoa'],
     );
 
@@ -187,7 +181,7 @@ class App {
     this.debug('running %o manifest construction...', this.id);
 
     // build the manifest from our plugins
-    const manifest = this.config.newConfig({id: 'manifest'});
+    const manifest = this.config.newConfig({ id: 'manifest' });
     // loop through plugins in reverse order and add them to the manifest config
     for (const [name, plugin] of Object.entries(this.plugins).reverse()) {
       // clone the manifest
@@ -198,7 +192,7 @@ class App {
       delete store.description;
       delete store.enabled;
       delete store.hidden;
-      manifest.add(name, {type: 'literal', store});
+      manifest.add(name, { type: 'literal', store });
     }
 
     // set
@@ -224,12 +218,12 @@ class App {
     this.debug('running %o plugin discovery...', this.name);
 
     // do the discovery
-    const {disabled, enabled, invalids} = require('../utils/get-plugins')(
+    const { disabled, enabled, invalids } = require('../utils/get-plugins')(
       [
         // these are plugins translated into single dir depth 0 searches
         ...Object.entries(this.appConfig.get('plugins'))
-          .map(([name, dir]) => ({dir, depth: 0}))
-          .filter(plugin => fs.existsSync(plugin.dir)),
+          .map(([, dir]) => ({ dir, depth: 0 }))
+          .filter((plugin) => fs.existsSync(plugin.dir)),
         // these are just the plugin dirs to search
         ...this.appConfig.get('plugin-dirs'),
       ],
@@ -247,7 +241,7 @@ class App {
     );
 
     // rebase on any passed in plugins
-    const plugins = {...this._plugins, ...enabled};
+    const plugins = { ...this._plugins, ...enabled };
 
     // set things
     this.#_cache.set('plugins.disabled', disabled);
@@ -259,7 +253,7 @@ class App {
   }
 
   getRegistry() {
-    return this.manifest.get('registry', {decode: false, encode: false});
+    return this.manifest.get('registry', { decode: false, encode: false });
   }
 
   // public wrapper for reint
@@ -291,8 +285,8 @@ class App {
   }
 
   async runHook(event, data) {
-    return require('../utils/run-hook')(event, data, this.hooks, {app: this}, this.debug);
+    return require('../utils/run-hook')(event, data, this.hooks, { app: this }, this.debug);
   }
 }
 
-module.exports = App;
+export default App;
