@@ -1,19 +1,19 @@
-'use strict';
+import os from 'node:os';
+import path from 'node:path';
+import { semver } from 'bun';
 
-const fs = require('fs-extra');
-const has = require('lodash/has');
-const isClass = require('is-class');
-const makeError = require('../utils/make-error');
-const merge = require('../utils/merge');
-const os = require('os');
-const path = require('path');
-const parsePkgName = require('../utils/parse-package-name');
-const read = require('../utils/read-file');
-const semver = require('semver');
-const write = require('../utils/write-file');
+import fs from 'fs-extra';
 
-const {nanoid} = require('nanoid');
-const {extract, manifest, packument} = require('pacote');
+import has from 'lodash-es/has';
+import isClass from 'is-class';
+import { extract, manifest, packument } from 'pacote';
+
+import generateId from '../utils/generate-id.js';
+import makeError from '../utils/make-error.js';
+import merge from '../utils/merge.js';
+import parsePkgName from '../utils/parse-package-name.js';
+import read from '../utils/read-file.js';
+import write from '../utils/write-file.js';
 
 /**
  *
@@ -29,32 +29,27 @@ class Plugin {
   /*
    * fetches a plugin from a registry/git repo
    */
-  static async fetch(plugin, {
-    dest = os.tmpdir(),
-    installer = Plugin.installer,
-    config = Plugin.config,
-    type = 'app',
-  } = {}) {
+  static async fetch(plugin, { dest = os.tmpdir(), installer = Plugin.installer, config = Plugin.config, type = 'app' } = {}) {
     // parse the package name
     const pkg = parsePkgName(plugin);
     // get the info so we can determine whether this is a lando package or not
-    const info = await Plugin.info(pkg.raw, {config});
+    const info = await Plugin.info(pkg.raw, { config });
 
     // update dest with name and compute the package.json location
     dest = path.join(dest, info.name);
     const pjson = path.join(dest, 'package.json');
 
     // make sure we have a place to extract the plugin
-    const tmp = path.join(os.tmpdir(), nanoid());
-    fs.mkdirSync(tmp, {recursive: true});
+    const tmp = path.join(os.tmpdir(), generateId());
+    fs.mkdirSync(tmp, { recursive: true });
 
     // try to extract the plugin
-    const {resolved} = await extract(pkg.raw, tmp, merge({Arborist: require('@npmcli/arborist')}, [config]));
+    const { resolved } = await extract(pkg.raw, tmp, merge({ Arborist: require('@npmcli/arborist') }, [config]));
     Plugin.debug('extracted plugin %o to %o from %o using %o', info._id, tmp, resolved, config);
 
     // if we get this far then we can safely move the plugin to dest
-    fs.rmSync(dest, {recursive: true, force: true});
-    fs.mkdirSync(dest, {recursive: true});
+    fs.rmSync(dest, { recursive: true, force: true });
+    fs.mkdirSync(dest, { recursive: true });
     fs.copySync(tmp, dest);
     Plugin.debug('moved plugin from %o to %o', tmp, dest);
 
@@ -65,22 +60,20 @@ class Plugin {
     }
 
     // return instantiated plugin
-    return new Plugin(dest, {installer, type});
+    return new Plugin(dest, { installer, type });
   }
 
   /*
    *
    * TBD
    */
-  static async info(plugin, {
-    config = Plugin.config,
-  } = {}) {
+  static async info(plugin, { config = Plugin.config } = {}) {
     // parse the plugin name
     const pkg = parsePkgName(plugin);
 
     // try to get info about the package
     try {
-      const info = await manifest(pkg.raw, merge({}, [config, {fullMetadata: true, preferOnline: true}]));
+      const info = await manifest(pkg.raw, merge({}, [config, { fullMetadata: true, preferOnline: true }]));
       Plugin.debug('retrieved plugin information for %o from %o using %o', pkg.raw, info._resolved, config);
 
       // if not a "lando plugin" then throw an error
@@ -93,12 +86,12 @@ class Plugin {
 
       return info;
 
-    // handle errors
+      // handle errors
     } catch (error) {
       // better 404 message
       if (error.statusCode === 404) error.message = `Could not find a plugin called ${pkg.raw}`;
       // throw
-      throw makeError({error});
+      throw makeError({ error });
     }
   }
 
@@ -108,9 +101,11 @@ class Plugin {
   static isValid(plugin) {
     // if we are looking at a plugin instance then
     if (isClass(plugin.constructor) && plugin.constructor.name === 'Plugin') {
-      return Object.keys(plugin.manifest).length > 0 ||
+      return (
+        Object.keys(plugin.manifest).length > 0 ||
         has(plugin.pjson, 'lando') ||
-        (plugin.pjson.keywords && plugin.pjson.keywords.includes('lando-plugin'));
+        (plugin.pjson.keywords && plugin.pjson.keywords.includes('lando-plugin'))
+      );
     }
 
     // if we get here assume its info and return true for needed conditions
@@ -126,15 +121,18 @@ class Plugin {
   /*
    * @TODO: scripts shoudl be moved into the engine constructor
    */
-  constructor(location, {
-    config = {},
-    debug = Plugin.debug,
-    id = Plugin.id || 'lando',
-    installer = Plugin.installer,
-    loadOpts = [],
-    type = 'app',
-    version = undefined,
-  } = {}) {
+  constructor(
+    location,
+    {
+      config = {},
+      debug = Plugin.debug,
+      // id = Plugin.id || 'lando',
+      installer = Plugin.installer,
+      loadOpts = [],
+      type = 'app',
+      // version = undefined,
+    } = {},
+  ) {
     // core props
     this.root = location;
     this.enabled = true;
@@ -151,7 +149,7 @@ class Plugin {
     // get the manifest
     debug.extend(this.pjson.name)('found plugin at %o', this.root);
     // get the manifest and normalize it based on root
-    this.manifest = require('../utils/normalize-manifest-paths')({...this.pjson.lando, ...this.#load(...loadOpts)}, this.root);
+    this.manifest = require('../utils/normalize-manifest-paths')({ ...this.pjson.lando, ...this.#load(...loadOpts) }, this.root);
 
     // if the manifest disables the plugin
     if (this.manifest.enabled === false || this.manifest.enabled === 0) this.enabled = false;
@@ -178,7 +176,7 @@ class Plugin {
       // remove the root
       ppsjons.pop();
       // paternity test
-      const pjsons = ppsjons.filter(pjson => fs.existsSync(pjson));
+      const pjsons = ppsjons.filter((pjson) => fs.existsSync(pjson));
       // if we have parents then use the closest
       if (pjsons.length > 0) this.version = require(pjsons[0]).version;
     }
@@ -201,7 +199,7 @@ class Plugin {
     }
 
     // log result
-    const status = {enabled: this.enabled, valid: this.isValid, installed: this.isInstalled};
+    const status = { enabled: this.enabled, valid: this.isValid, installed: this.isInstalled };
     this.debug('instantiated plugin with status %o', status);
   }
 
@@ -227,7 +225,7 @@ class Plugin {
 
     try {
       // get release data
-      const data = await packument(this.spec, {fullMetadata: true});
+      const data = await packument(this.spec, { fullMetadata: true });
       // build a list of highest available versions
       const havs = [data['dist-tags'].latest];
       // if we are looking at a non-standard channel then include that tag as well
@@ -240,19 +238,19 @@ class Plugin {
         this.debug('no update available on the %o channel, %o <= %o', channel, hv, this.version);
         return false;
 
-      // otherwise update is available
+        // otherwise update is available
       } else {
         this.debug('update available on the %o channel, %o <= %o, ', channel, hv, this.version);
         return true;
       }
 
-    // catch
-    // @TODO: what do we actually want to do here?
+      // catch
+      // @TODO: what do we actually want to do here?
     } catch (error) {
       // better 404 message
-      if (error.statusCode === 404) error.message = `Could not find a plugin called ${pkg.raw}`;
+      if (error.statusCode === 404) error.message = `Could not find a plugin called ${this.name}`;
       // throw
-      throw makeError({error});
+      throw makeError({ error });
     }
   }
 
@@ -261,7 +259,7 @@ class Plugin {
    * across different package types
    */
   async info() {
-    const info = await manifest(this.location, {fullMetadata: true});
+    const info = await manifest(this.location, { fullMetadata: true });
     this.debug('retrieved plugin information for %o from %o', this.spec, info._resolved);
     return info;
   }
@@ -270,7 +268,7 @@ class Plugin {
    * Install a plugin.
    * ideally we do not have to do this because the pacakge is fully contained in the tarball
    */
-  async install({installer = this.installer || Plugin.installer} = {}) {
+  async install({ installer = this.installer || Plugin.installer } = {}) {
     // try teh install
     await installer.installPlugin(this.root, this.config.installer);
     // if we get here then we can update isInstalled
@@ -281,7 +279,7 @@ class Plugin {
    * Remove a plugin.
    */
   remove() {
-    return fs.rmSync(this.root, {recursive: true, force: true});
+    return fs.rmSync(this.root, { recursive: true, force: true });
   }
 
   /*
@@ -302,7 +300,6 @@ class Plugin {
   //     const hv = semver.rsort(havs)[0];
   //     // @TODO: calculate the destination?
 
-
   //   // catch
   //   // @TODO: what do we actually want to do here?
   //   } catch (error) {
@@ -314,4 +311,4 @@ class Plugin {
   // }
 }
 
-module.exports = Plugin;
+export default Plugin;
