@@ -1,9 +1,7 @@
-'use strict';
-
-module.exports = cli => ({
+export default (cli) => ({
   command: 'registry',
   describe: 'displays the registry for this context',
-  run: async (options, {ctx}) => {
+  run: async (options, { ctx }) => {
     cli.log(ctx.getRegistry());
   },
 });
