@@ -1,13 +1,11 @@
-'use strict';
-
-const _ = require('lodash');
+import _ from 'lodash-es';
 
 // Helper to filter services
 const filterServices = (service, services = []) => {
   return !_.isEmpty(services) ? _.includes(services, service) : true;
 };
 
-module.exports = cli => ({
+export default (cli, lando) => ({
   command: 'info',
   describe: 'Prints info about your app',
   // options: _.merge({}, cli.formatOptions(), {
@@ -23,23 +21,28 @@ module.exports = cli => ({
   //     array: true,
   //   },
   // }),
-  run: options => {
+  run: (options) => {
     // Try to get our app
     const app = lando.getApp(options._app.root);
     // Get services
-    app.opts = (!_.isEmpty(options.service)) ? {services: options.service} : {};
+    app.opts = !_.isEmpty(options.service) ? { services: options.service } : {};
     // Go deep if we need to
     if (app && options.deep) {
-      return app.init().then(() => lando.engine.list({project: app.project})
-        .filter(container => filterServices(container.service, options.service))
-        .each(container => lando.engine.scan(container)
-          .then(data => console.log(cli.formatData(data, options)))),
-        );
+      return app.init().then(() =>
+        lando.engine
+          .list({ project: app.project })
+          .filter((container) => filterServices(container.service, options.service))
+          .each((container) => lando.engine.scan(container).then((data) => console.log(cli.formatData(data, options)))),
+      );
     } else if (app && !options.deep) {
-      return app.init().then(() => console.log(cli.formatData(
-        _.filter(app.info, service => filterServices(service.service, options.service)),
-        options,
-      )));
+      return app.init().then(() =>
+        console.log(
+          cli.formatData(
+            _.filter(app.info, (service) => filterServices(service.service, options.service)),
+            options,
+          ),
+        ),
+      );
     }
   },
 });
