@@ -1,13 +1,24 @@
-'use strict';
+import camelCase from 'lodash-es/camelCase';
+import camelcaseKeys from 'camelcase-keys';
 
 // @TODO: throw error for nully values?
-module.exports = data => {
+export default function decode(data) {
   // if we have a nully value then just return
   if (data === null || data === undefined) return data;
   // if string then return
-  if (typeof data === 'string') return data.split('.').map(part => require('lodash/camelCase')(part)).join('.');
+  if (typeof data === 'string')
+    return data
+      .split('.')
+      .map((part) => camelCase(part))
+      .join('.');
   // if array then map and return
-  if (Array.isArray(data)) return data.map(prop => prop.split('.').map(part => require('lodash/camelCase')(part)).join('.'));
+  if (Array.isArray(data))
+    return data.map((prop) =>
+      prop
+        .split('.')
+        .map((part) => camelCase(part))
+        .join('.'),
+    );
   // else assume object and return
-  return require('camelcase-keys')(data, {deep: true});
-};
+  return camelcaseKeys(data, { deep: true });
+}
