@@ -1,20 +1,19 @@
-'use strict';
+import isObject from 'lodash-es/isPlainObject.js';
+import get from 'lodash-es/get.js';
+import set from 'lodash-es/set.js';
 
-const isObject = require('lodash/isPlainObject');
-const get = require('lodash/get');
-const getKeys = require('./get-object-keys');
-const getSize = require('./get-size');
-const set = require('lodash/set');
+import getKeys from './get-object-keys.js';
+import getSize from './get-size.js';
 
 /*
  * TBD
  */
-module.exports = (data, {depth = 1, sizes = {}} = {}) => {
+export default function getObjectSize(data, { depth = 1, sizes = {} } = {}) {
   // @TODO: implement depth?
   // @TODO: throw error if data is not object?
   if (!isObject(data)) return {};
   // set all the sizes
-  for (const path of getKeys(data, {depth, expandArrays: false})) set(sizes, path, getSize(get(data, path)));
+  for (const path of getKeys(data, { depth, expandArrays: false })) set(sizes, path, getSize(get(data, path)));
   // return sizes
   return sizes;
-};
+}
