@@ -1,17 +1,24 @@
-'use strict';
+import path from 'node:path';
 
-const path = require('path');
+import getEnvironment from './get-environment.js';
+import getPlatform from './get-platform.js';
+
+const env = getEnvironment();
 
 /*
  * TBD
  */
-module.exports = (id = 'lando') => {
-  switch (process.platform) {
+
+// @TODO: this needs to use getPlatform()?
+
+export default function getSystemDataDir(id = 'devtool') {
+  switch (getPlatform()) {
     case 'darwin':
       return path.join('/', 'Library', 'Application Support', `${id[0].toUpperCase()}${id.slice(1).toLowerCase()}`);
     case 'linux':
+    case 'wsl':
       return path.join('/', 'srv', `${id.toLowerCase()}`);
     case 'win32':
-      return path.join(process.env.PROGRAMDATA || process.env.ProgramData, `${id[0].toUpperCase()}${id.slice(1).toLowerCase()}`);
+      return path.join(env.PROGRAMDATA ?? env.ProgramData, `${id[0].toUpperCase()}${id.slice(1).toLowerCase()}`);
   }
-};
+}
