@@ -1,11 +1,9 @@
-'use strict';
-
-const isObject = require('lodash/isPlainObject');
-const merge = require('lodash/mergeWith');
-const mergeArrays = require('./merge-arrays');
+import isObject from 'lodash-es/isPlainObject.js';
+import lomerge from 'lodash-es/merge.js';
+import mergeArrays from './merge-arrays.js';
 
 // @TODO: error handling
-module.exports = (object, sources, ams = ['merge:id', 'replace']) => {
+export default function merge(object, sources, ams = ['merge:id', 'replace']) {
   // if sources is not an array then make it so
   if (!Array.isArray(sources)) sources = [sources];
 
@@ -17,7 +15,7 @@ module.exports = (object, sources, ams = ['merge:id', 'replace']) => {
   const fallback = ams[1] || 'replace';
 
   // then return the merge
-  return merge(object, ...sources, (objValue, srcValue) => {
+  return lomerge(object, ...sources, (objValue, srcValue) => {
     // if not an arrayjust proceed normally
     if (!Array.isArray(objValue)) return undefined;
     // if first strategy is replace then also proceed normally
@@ -25,11 +23,11 @@ module.exports = (object, sources, ams = ['merge:id', 'replace']) => {
     // if first strategy is merge and some part of objvalue has an object then try to merge with by
     if (first === 'merge') {
       // if mergable object detected in array then proceed
-      if (objValue.some(element => isObject(element))) return mergeArrays(objValue, srcValue, `merge:${by}`);
+      if (objValue.some((element) => isObject(element))) return mergeArrays(objValue, srcValue, `merge:${by}`);
       // otherwise return the fallback
       return mergeArrays(objValue, srcValue, fallback);
     }
     // if we get here then just pass it through to mergeArrays
     return mergeArrays(objValue, srcValue, first);
   });
-};
+}
