@@ -1,11 +1,11 @@
-'use strict';
+import npmPkgArg from 'npm-package-arg';
 
 /*
  * TBD
  */
-module.exports = (plugin, {defaultTag = 'stable'} = {}) => {
+export default function parsePkgName(plugin /* { /*defaultTag = 'stable' } = {}*/) {
   // parse the plugin
-  const result = require('npm-package-arg')(plugin);
+  const result = npmPkgArg(plugin);
 
   // add a package property
   result.package = result.scope ? result.name.replace(`${result.scope}/`, '') : result.name;
@@ -15,4 +15,4 @@ module.exports = (plugin, {defaultTag = 'stable'} = {}) => {
 
   // return
   return result;
-};
+}
