@@ -1,7 +1,5 @@
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
+import fs from 'node:fs';
+import path from 'node:path';
 
 /*
  * Like readdirsync but returns absolute paths and only directories
@@ -9,6 +7,9 @@ const path = require('path');
  * NOTE: we are using something like this mostly just for bootstrapping eg
  * we want it to have minimal deps eg no LODASH or GLOB
  */
-module.exports = dir => fs.readdirSync(dir)
-.map(file => path.join(dir, file))
-.filter(file => fs.statSync(file).isDirectory());
+export default function readDirAbsolute(dir) {
+  return fs
+    .readdirSync(dir)
+    .map((file) => path.join(dir, file))
+    .filter((file) => fs.statSync(file).isDirectory());
+}
