@@ -1,42 +1,43 @@
-'use strict';
-
-const ModuleLoader = require('../lib/module-loader');
+import createDebug from '../lib/debug.js';
+import load from './module-loader.js';
 
 /*
  * TBD
  */
-module.exports = async (event, data, hooks = [], context = {}, debug = require('../lib/debug')('@lando/core'), errorHandler) => {
+export default async function runHook(event, data, hooks = [], context = {}, debug = createDebug('devtool:run-hooks'), errorHandler) {
   // debugger
-  debug = debug.extend(`run-hook:${event}`);
+  debug = debug.extend(`hook:${event}`);
   // collect successes and failures
-  const final = {successes: [], failures: []};
+  const final = { successes: [], failures: [] };
   // just helpful message to indicate we have started a cli hook
   debug('start %o hook', event);
-  // get hooks
-  const groups = hooks[event] || [];
+  // get targets
+  const groups = hooks[event] ?? [];
+  const targets = groups.map((group) => group.target ?? group);
+
   // generate promises from hooks
-  const promises = groups.map(async group => {
+  const promises = targets.map(async (target) => {
     // if this is not a group then make it a group of one
-    if (typeof group === 'string') group = [group];
+    if (typeof target === 'string') target = [target];
 
     // @NOTE: i dont think "this" works in async functions?
-    const ctx = {...context, debug: debug.contract(-2).extend(`hook:${event}`)};
+    const ctx = { ...context, debug: debug.contract(-2).extend(`hook:${event}`) };
 
     // loop through stringy existy runners and try to get results
-    for (const runner of group.filter(runner => runner && typeof runner === 'string')) {
+    for (const runner of target.filter((runner) => runner && typeof runner === 'string')) {
       try {
-        const {isESM, module, filePath} = await ModuleLoader.load(runner);
-        debug('start %s %o', isESM ? '(import)' : '(require)', filePath);
+        const { isESM, module, file } = await load(runner);
+        debug('start %s %o', isESM ? '(import)' : '(require)', file);
         // try to get result
         // @TODO: eventually cli/config are specific to this implementation so they need to be
         // taken out and passed in when we move to the util
-        const result = await module.call(ctx, {...data, ...ctx});
-        final.successes.push({group, result});
-        debug('done %o', filePath);
+        const result = await module.call(ctx, { ...data, ...ctx });
+        final.successes.push({ target, result });
+        debug('done %o', file);
 
-      // put errors here?
+        // put errors here?
       } catch (error) {
-        final.failures.push({group, error});
+        final.failures.push({ target, error });
         debug(error);
       }
     }
@@ -51,4 +52,4 @@ module.exports = async (event, data, hooks = [], context = {}, debug = require('
 
   // otherwise
   return final;
-};
+}
