@@ -1,19 +1,13 @@
-'use strict';
-
 /*
  * Attempts to produce a standardized error object
  */
-module.exports = ({
-  all,
-  args,
-  command,
-  stdout,
-  stderr,
-}) => ({
-  command,
-  args,
-  exitCode: 0,
-  stdout,
-  stderr,
-  all,
-});
+export default function makeSuccess({ all, args, command, stdout, stderr }) {
+  return {
+    command,
+    args,
+    exitCode: 0,
+    stdout,
+    stderr,
+    all,
+  };
+}
